requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
duckduckgo-search
python-dotenv
numpy